Adds caching to manga routes
Implements caching for manga routes using FastAPI-Cache2 and Redis to improve response times and reduce load on the backend.

Adds a lifespan context manager to initialize the Redis backend for FastAPI-Cache2.

Applies caching with a 360-second expiration time to the recent updates, popular, manga details, chapter details, and genre routes.